Refactor utility imports and consolidate file utilities into dedicated modules
