+++ py_code/tools/code_editing/__init__.py
@@ -1,8 +1,9 @@
 """Code editing tools package."""
 
+from ..utils.code_utils import read_code_from_path
 from .file_operations import delete_file, move_file_or_folder
 from .list_dir import list_dir
-from .mcp_tools import get_server_info, read_code_from_path
+from .mcp_tools import get_server_info
 from .read_file import read_file
 
 __all__ = ["list_dir", "read_file", "read_code_from_path", "get_server_info", "move_file_or_folder", "delete_file"]
