+++ py_code/tools/code_editing/list_dir.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, Optional
 
 from fastmcp import Context
 
-from ..utils import normalize_path
+from ..utils.file_utils import normalize_path
 
 
 def list_dir(path: str, ctx: Optional[Context] = None) -> Dict[str, Any]:
