+++ py_code/tools/utils/__init__.py
@@ -1,0 +1,13 @@
+"""Utilities for MCP tools."""
+
+from .code_utils import read_code_from_path
+from .file_utils import filter_binary_files, get_file_contents, get_lines_from_file, is_binary_file, normalize_path
+
+__all__ = [
+    "normalize_path",
+    "is_binary_file",
+    "filter_binary_files",
+    "get_file_contents",
+    "get_lines_from_file",
+    "read_code_from_path",
+]
