+++ py_code/tools/utils/file_utils/path_utils.py
@@ -1,0 +1,16 @@
+"""Path utility functions for MCP tools."""
+
+import os
+
+
+def normalize_path(path: str) -> str:
+    """Normalize a file path for consistent handling.
+
+    Args:
+        path: The path to normalize
+
+    Returns:
+        The normalized file path
+
+    """
+    return os.path.normpath(path)
